REFAC: Refactored reminders into a manager
