+++ eslint.config.mjs
@@ -4,8 +4,6 @@ import simpleImportSort from "eslint-plugin-simple-import-sort";
 import eslintPluginUnicorn from "eslint-plugin-unicorn";
 import tseslint from "typescript-eslint";
 
-// console.log(...obsidianmd.configs.recommended);
-
 export default tseslint.config(
     eslint.configs.recommended,
     ...tseslint.configs.recommended,
