+++ package.json
@@ -35,9 +35,9 @@
         "concurrently": "^9.2.1",
         "esbuild": "^0.28.0",
         "eslint": "^10.2.1",
+        "eslint-plugin-obsidianmd": "^0.3.0",
         "eslint-plugin-simple-import-sort": "^12.1.1",
         "eslint-plugin-unicorn": "^64.0.0",
-        "eslint-plugin-obsidianmd": "^0.3.0",
         "jest": "^30.3.0",
         "jest-environment-jsdom": "^30.3.0",
         "jest-expect-message": "^1.1.3",
@@ -52,6 +52,7 @@
         "vhtml": "^2.2.0"
     },
     "dependencies": {
+        "@jest/globals": "^30.4.1",
         "chart.js": "^4.5.1",
         "clozecraft": "^0.4.0",
         "gridjs": "^6.2.0",
