+++ src/main.ts
@@ -18,15 +18,15 @@ import { LocaleManagerInstance } from "src/lang/locale-manager";
 import { NextNoteReviewHandler } from "src/note/next-note-review-handler";
 import { Note } from "src/note/note";
 import { NoteReviewQueue } from "src/note/note-review-queue";
-import { RepItemState } from "src/scheduling/algorithms/base/repetition-item";
 import { SRAlgorithm } from "src/scheduling/algorithms/base/sr-algorithm";
 import {
     FlashcardReviewMode,
     FlashcardReviewSequencer,
     IFlashcardReviewSequencer,
 } from "src/scheduling/flashcard-review-sequencer";
+import { ReminderManager } from "src/scheduling/reminder-manager";
 import { REVIEW_QUEUE_VIEW_TYPE } from "src/ui/obsidian-ui-components/item-views/review-queue-list-view";
-import { UIManager, UIState } from "src/ui/ui-manager";
+import { UIManager } from "src/ui/ui-manager";
 import { TextDirection } from "src/utils/strings";
 
 export default class SRPlugin extends Plugin {
@@ -35,9 +35,8 @@ export default class SRPlugin extends Plugin {
 
     private _nextNoteReviewHandler: NextNoteReviewHandler | null = null;
     private _commandManager: CommandManager | null = null;
+    private _reminderManager: ReminderManager | null = null;
     public isInitialized: boolean = false;
-    private reviewReminderTimer: number | null = null;
-    private isReviewReminderChecking: boolean = false;
 
     async onload(): Promise<void> {
         try {
@@ -79,9 +78,10 @@ export default class SRPlugin extends Plugin {
 
                 await this.uiManager.onLayoutReady();
                 this.commandManager.onLayoutReady();
+                this._reminderManager = new ReminderManager(this, this.uiManager, this.dataManager);
 
                 this.isInitialized = true;
-                this.restartReviewReminders();
+                this._reminderManager.restartReviewReminders();
             });
         } catch (error) {
             if (error instanceof PluginDataError || error instanceof Error) {
@@ -95,6 +95,11 @@ export default class SRPlugin extends Plugin {
                 DebugLoggerInstance.getInstance().log("SRPlugin: Error in onLoad", "error");
             }
         }
+    }
+
+    get reminderManager(): ReminderManager {
+        if (this._reminderManager === null) throw new Error("Reminder manager not initialized!!!");
+        return this._reminderManager;
     }
 
     get uiManager(): UIManager {
@@ -139,15 +144,10 @@ export default class SRPlugin extends Plugin {
     }
 
     onunload(): void {
-        this.stopReviewReminders();
+        this.reminderManager.stopReviewReminders();
         this.app.workspace.getLeavesOfType(REVIEW_QUEUE_VIEW_TYPE).forEach((leaf) => leaf.detach());
         this.uiManager.destroy();
         this.commandManager.onunload();
-    }
-
-    public restartReviewReminders() {
-        this.stopReviewReminders();
-        this.startReviewReminders();
     }
 
     public addCustomHotkeys() {
@@ -216,138 +216,6 @@ export default class SRPlugin extends Plugin {
         if (this.dataManager.data.settings.enableNoteReviewPaneOnStartup) {
             this.uiManager.sidebarManager.redraw();
         }
-    }
-
-    /**
-     * Starts the periodic review-reminder loop once the plugin is fully initialized.
-     *
-     * The reminder feature depends on loaded settings, synced review data, and a ready UI.
-     * Running it earlier risks duplicate startup checks or reminders based on stale state.
-     */
-    private startReviewReminders() {
-        if (!this.isInitialized || !this.dataManager.data.settings.enableReviewReminders) {
-            return;
-        }
-
-        if (this.dataManager.data.settings.reviewReminderCheckOnStartup) {
-            // Startup checking is a one-shot pass. The periodic interval below remains the only
-            // long-lived scheduler so reminder timing stays predictable.
-            void this.checkReviewReminders();
-        }
-
-        const intervalMinutes = Math.min(
-            Math.max(this.dataManager.data.settings.reviewReminderIntervalMinutes, 1),
-            1440,
-        );
-        this.reviewReminderTimer = window.setInterval(
-            () => {
-                void this.checkReviewReminders();
-            },
-            intervalMinutes * 60 * 1000,
-        );
-    }
-
-    /**
-     * Stops the reminder interval and clears the in-flight guard.
-     *
-     * The guard is reset here as well so that disabling and immediately re-enabling reminders
-     * never leaves the scheduler stuck in a "currently checking" state.
-     */
-    private stopReviewReminders() {
-        if (this.reviewReminderTimer !== null) {
-            window.clearInterval(this.reviewReminderTimer);
-            this.reviewReminderTimer = null;
-        }
-        this.isReviewReminderChecking = false;
-    }
-
-    /**
-     * Performs a single reminder check.
-     *
-     * We intentionally short-circuit when SR is already open, the user is editing text, or a
-     * sync is already running. In those states, a reminder would be either redundant or
-     * disruptive, and auto-opening review would compete with the user's current context.
-     */
-    private async checkReviewReminders() {
-        if (
-            !this.isInitialized ||
-            !this.isDataManagerLoaded() ||
-            !this.dataManager.isOsrCoreLoaded() ||
-            this.isReviewReminderChecking ||
-            !this.dataManager.data.settings.enableReviewReminders
-        ) {
-            return;
-        }
-
-        if (this.uiManager.uiState !== UIState.Closed || this.uiManager.isSRInFocus) {
-            return;
-        }
-
-        if (this.isUserEditingText() || this.dataManager.syncLock) {
-            return;
-        }
-
-        this.isReviewReminderChecking = true;
-        try {
-            await this.dataManager.sync();
-
-            const remainingDeckTree = this.dataManager.osrCore.remainingDeckTree;
-            if (remainingDeckTree === null) {
-                return;
-            }
-
-            const reviewCardCount = remainingDeckTree.getDistinctRepItemCount(
-                RepItemState.AnyItem,
-                true,
-            );
-            if (
-                reviewCardCount <= 0 ||
-                this.uiManager.uiState !== UIState.Closed ||
-                this.uiManager.isSRInFocus ||
-                this.isUserEditingText()
-            ) {
-                return;
-            }
-
-            await this.uiManager.notifyReviewReminder();
-
-            if (this.dataManager.data.settings.reviewReminderAutoOpen) {
-                // Attention and navigation are separate concerns: users can keep the reminder
-                // signal without consenting to review auto-open, or enable both together.
-                await this.uiManager.openDeckContainer(FlashcardReviewMode.Review);
-            }
-        } finally {
-            this.isReviewReminderChecking = false;
-        }
-    }
-
-    /**
-     * Detects whether the user is actively typing in Obsidian.
-     *
-     * The reminder flow uses this to avoid stealing focus or opening review while the user is in
-     * a text field, contenteditable surface, or CodeMirror-backed editor.
-     */
-    private isUserEditingText(): boolean {
-        const activeElement = activeDocument.activeElement;
-        if (activeElement === null) {
-            return false;
-        }
-
-        if (
-            activeElement.instanceOf(HTMLTextAreaElement) ||
-            activeElement.instanceOf(HTMLInputElement)
-        ) {
-            return true;
-        }
-
-        if (activeElement.instanceOf(HTMLElement)) {
-            return (
-                activeElement.isContentEditable ||
-                activeElement.closest(".cm-editor, .markdown-source-view") !== null
-            );
-        }
-
-        return false;
     }
 
     private static createDeckTreeIterator(settings: SRSettings): IDeckTreeIterator {
